requirements.txt
asgiref==3.5.0
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
